+++ src/components/base/BaseInput/BaseInputDefault.vue
@@ -38,6 +38,8 @@
 							:value="vmodel"
 							:disabled="disabled"
 							:class="classes"
+							:min="appliedType === 'date' ? '1900-01-01' : ''"
+							:max="appliedType === 'date' ? '2999-12-31' : ''"
 							@input="handleInput"
 							@focus="hasFocus = true"
 							@blur="hasFocus = false"
@@ -93,6 +95,8 @@ export const supportedTypes = [
 	"textarea",
 	"url",
 	"number",
+	"date",
+	"time",
 ];
 
 export default {
@@ -180,13 +184,13 @@ export default {
 	&:focus-within,
 	&:hover:not(.disabled) {
 		.label {
-			color: var(--color-accent);
+			color: var(--color-primary);
 		}
 		.help {
 			visibility: visible;
 		}
 		.visible {
-			fill: var(--color-accent);
+			fill: var(--color-primary);
 		}
 	}
 }
@@ -197,7 +201,7 @@ export default {
 
 	&:focus-within,
 	&:hover:not(.disabled) {
-		border-bottom-color: var(--color-accent);
+		border-bottom: var(--border-width-bold) solid var(--color-primary);
 		outline: none;
 	}
 	&.error {