+++ deploy/deploy.sh
@@ -33,6 +33,13 @@ inform_staging() {
   fi
 }
 
+inform_hotfix() {
+  if [[ "$TRAVIS_EVENT_TYPE" != "cron" ]]
+  then
+    curl -X POST -H 'Content-Type: application/json' --data '{"text":":boom: Das Hotfix-'$1'-System wurde aktualisiert: HPI Schul-Cloud Nuxt-Client! https://hotfix'$1'.schul-cloud.dev/nuxtversion (Dockertag: '$DOCKERTAG')"}' $WEBHOOK_URL_CHAT
+  fi
+}
+
 deploy(){
 	SYSTEM=$1 # [staging, test, demo]
 
@@ -43,6 +50,7 @@ deploy(){
 	COMPOSE_SRC=$5 # name of the docker-compose file which should be used as.
 	COMPOSE_TARGET=$6 # name as which the compose file should be pushed to the server (auto prefixed with "docker-compose-")
 	STACK_NAME=$7 # swarm stack name
+	TLD=${8:=org}  # If variable not set or null, set it to default.
 
 	echo "deploy " $DOCKER_IMAGE ":" $DOCKER_TAG " to " $SYSTEM " as " $DOCKER_SERVICE_NAME
 	echo "COMPOSEFILE: " $COMPOSE_SRC " => " $COMPOSE_TARGET
@@ -50,12 +58,15 @@ deploy(){
 	# generate new compose file
 	eval "echo \"$( cat $COMPOSE_SRC )\"" > docker-compose-$COMPOSE_TARGET
 
+	# info: we don't need the two lines for deployment because they use the false docker-compose file and overwrites the config which created by devops, this results from the fact that the wrong csp configuration is used and the environment variables are missing if they are not explicitly included in the template. (pr 1158)
+
 	# deploy new compose file
-	scp -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null -i travis_rsa docker-compose-$COMPOSE_TARGET linux@$SYSTEM.schul-cloud.org:~
-	ssh -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null -i travis_rsa linux@$SYSTEM.schul-cloud.org /usr/bin/docker stack deploy -c /home/linux/docker-compose-$COMPOSE_TARGET $STACK_NAME
+	# scp -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null -i travis_rsa docker-compose-$COMPOSE_TARGET linux@$SYSTEM.schul-cloud.org:~
+	# ssh -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null -i travis_rsa linux@$SYSTEM.schul-cloud.org /usr/bin/docker stack deploy -c /home/linux/docker-compose-$COMPOSE_TARGET $STACK_NAME
 
 	# deploy new dockerfile
-	ssh -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null -i travis_rsa linux@$SYSTEM.schul-cloud.org /usr/bin/docker service update --force --image schulcloud/schulcloud-$DOCKER_IMAGE:$DOCKER_TAG $DOCKER_SERVICE_NAME
+	echo "Attempting to update Docker service $DOCKER_SERVICE_NAME..."
+	ssh -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null -i travis_rsa linux@$SYSTEM.schul-cloud.$TLD /usr/bin/docker service update --force --image schulcloud/schulcloud-$DOCKER_IMAGE:$DOCKER_TAG $DOCKER_SERVICE_NAME
 }
 
 # ----------------
@@ -81,11 +92,14 @@ fi
 case "$TRAVIS_BRANCH" in
 
 	master)
+		# If an event occurs on branch master call inform without deployment
+		echo "Event detected on branch master. Informing team. No deployment on master."
 		inform_live $PROJECT
 		;;
 
 	develop)
-		echo "develop"
+		# If an event occurs on branch develop deploy to test
+		echo "Event detected on branch develop. Attempting to deploy to development (test) environment..."
 		case "$PROJECT" in
 			client)
 				# deploy $SYSTEM $DOCKERFILE $DOCKERTAG $DOCKER_SERVICENAME $COMPOSE_DUMMY $COMPOSE_FILE $COMPOSE_SERVICENAME
@@ -97,10 +111,14 @@ case "$TRAVIS_BRANCH" in
 			vuepress)
 				deploy "test" "nuxt-vuepress" $DOCKERTAG "test-schul-cloud_vuepress" "compose-vuepress.dummy" "nuxt-vuepress.yml" "test-schul-cloud"
 			;;
+			*)
+				echo "$PROJECT does not match one of \"client\", \"storybook\" or \"vuepress\". Deployment will be skipped."
+			;;
 		esac
 		;;
-	release* | hotfix*)
-		echo "release/hotfix"
+	release*)
+		# If an event occurs on branch release* deploy to staging
+		echo "Event detected on branch release*. Attempting to deploy to staging environment..."
 		case "$PROJECT" in
 			client)
 				# TODO deploy with themes
@@ -109,6 +127,7 @@ case "$TRAVIS_BRANCH" in
 				# deploy "staging" "nuxt-client" $DOCKERTAG "staging-schul-cloud_nuxtclient" "compose-client_n21.dummy" "nuxt-client_n21.yml" "staging-schul-cloud"
 				# deploy "staging" "nuxt-client" $DOCKERTAG "staging-schul-cloud_nuxtclient" "compose-client_open.dummy" "nuxt-client_open.yml" "staging-schul-cloud"
 				# deploy "staging" "nuxt-client" $DOCKERTAG "staging-schul-cloud_nuxtclient" "compose-client_thr.dummy" "nuxt-client_thr.yml" "staging-schul-cloud"
+				# deploy "staging" "nuxt-client" $DOCKERTAG "staging-schul-cloud_nuxtclient" "compose-client_int.dummy" "nuxt-client_int.yml" "staging-schul-cloud"
 			;;
 			storybook)
 				deploy "staging" "nuxt-storybook" $DOCKERTAG "staging_storybook" "compose-storybook.dummy" "nuxt-storybook.yml" "staging"
@@ -116,7 +135,44 @@ case "$TRAVIS_BRANCH" in
 			vuepress)
 				deploy "staging" "nuxt-vuepress" $DOCKERTAG "staging_vuepress" "compose-vuepress.dummy" "nuxt-vuepress.yml" "staging"
 			;;
+			*)
+				echo "$PROJECT does not match one of \"client\", \"storybook\" or \"vuepress\". Deployment will be skipped."
+			;;
 		esac
+		;;
+	hotfix*)
+		# If an event occurs on branch hotfix* parse team id 
+		# and deploy to according hotfix environment
+		TEAM="$(cut -d'/' -f2 <<< $TRAVIS_BRANCH)"
+		if [[ "$TEAM" -gt 0 && "$TEAM" -lt 8 ]]; then
+			echo "Event detected on branch hotfix/$TEAM/... . Attempting to deploy to hotfix environment $TEAM, project $PROJECT..."
+			inform_hotfix $TEAM
+			case "$PROJECT" in
+				client)
+					# TODO deploy with themes
+					deploy "hotfix${TEAM}" "nuxt-client" $DOCKERTAG "hotfix${TEAM}_nuxtclient" "compose-client_default.dummy" "nuxt-client_default.yml" "hotfix${TEAM}_nuxtclient" "dev"
+					# deploy "staging" "nuxt-client" $DOCKERTAG "staging-schul-cloud_nuxtclient" "compose-client_brb.dummy" "nuxt-client_brb.yml" "staging-schul-cloud"
+					# deploy "staging" "nuxt-client" $DOCKERTAG "staging-schul-cloud_nuxtclient" "compose-client_n21.dummy" "nuxt-client_n21.yml" "staging-schul-cloud"
+					# deploy "staging" "nuxt-client" $DOCKERTAG "staging-schul-cloud_nuxtclient" "compose-client_open.dummy" "nuxt-client_open.yml" "staging-schul-cloud"
+					# deploy "staging" "nuxt-client" $DOCKERTAG "staging-schul-cloud_nuxtclient" "compose-client_thr.dummy" "nuxt-client_thr.yml" "staging-schul-cloud"
+				;;
+				storybook)
+					deploy "hotfix${TEAM}" "nuxt-storybook" $DOCKERTAG "hotfix${TEAM}_nuxtclient" "compose-storybook.dummy" "nuxt-storybook.yml" "hotfix${TEAM}_nuxtclient" "dev"
+				;;
+				vuepress)
+					deploy "hotfix${TEAM}" "nuxt-vuepress" $DOCKERTAG "hotfix${TEAM}_nuxtclient" "compose-vuepress.dummy" "nuxt-vuepress.yml" "hotfix${TEAM}_nuxtclient" "dev"
+				;;
+				*)
+					echo "$PROJECT does not match one of \"client\", \"storybook\" or \"vuepress\". Deployment will be skipped."
+				;;
+			esac
+		else
+			echo "Event detected on branch hotfix*. However, branch name pattern does not match requirements to deploy. Expected hotfix/<team_number>/XX.XX.XX but got $TRAVIS_BRANCH"
+		fi
+		;;
+	*)
+		# If no condition is met, nothing will be deployed.
+  		echo "Event detected which does not meet any conditions. Deployment will be skipped."
 		;;
 esac