+++ src/components/organisms/ContentCard.vue
@@ -29,7 +29,9 @@
 								/>
 							</div>
 						</div>
-						<h6 class="content__title">{{ resource.name }}</h6>
+						<h6 class="content__title">
+							{{ resource.title || resource.name }}
+						</h6>
 					</div>
 				</template>
 			</base-link>
@@ -127,7 +129,6 @@ export default {
 .content {
 	display: flex;
 	flex-direction: column;
-	min-height: 300px;
 	&__img {
 		&-thumbnail {
 			width: 100%;
@@ -151,7 +152,7 @@ export default {
 			top: 50%;
 			left: 50%;
 			z-index: var(--layer-page);
-			font-size: var(--space-xl-3);
+			font-size: var(--space-xl-3) !important;
 			border-radius: var(--radius-round);
 			box-shadow: var(--shadow-m);
 			opacity: 0.8;
@@ -167,7 +168,7 @@ export default {
 		}
 	}
 	&__title {
-		min-height: 62px;
+		height: calc(var(--heading-6) * var(--line-height-sm) * 3);
 		margin: var(--space-xs) var(--space-sm);
 		color: var(--color-tertiary);